requirements.txt
numpy
torch>=1.0
torchvision
matplotlib
terminaltables
pillow
tqdm
flask